update controller holiday, schedule, paid leave. + view schedule

+++ app/Http/Controllers/HolidayController.php
@@ -87,49 +87,70 @@ class HolidayController extends Controller
             $check_day = date('j', strtotime($check_days));
 
             $accept_paid_leave = DB::table('accepted_paid_leaves')
-            ->where('date', '=', $check_year.'-'.$month.'-'.$check_day)->get();
+            ->where('date', '=', $check_year.'-'.$month.'-'.$check_day)
+            ->leftJoin('master_users','accepted_paid_leaves.user_id','=','master_users.id')
+            ->select(
+                'accepted_paid_leaves.*',
+                'master_users.division_id as division_id'
+                )
+            ->get();
 
             foreach ($accept_paid_leave as $item_accept) {
-                DB::statement('UPDATE transaction_paid_leaves SET `days` = `days` - 1 WHERE `id` = '.$item_accept->paid_leave_id);
-                DB::statement('UPDATE master_users SET `yearly_leave_remaining` = `yearly_leave_remaining` + 1 WHERE `id` = '.$item_accept->user_id);
-                DB::table('accepted_paid_leaves')->where('id', '=', $item_accept->id)->delete();
+                if ($item_accept->division_id != 5) {
+                    DB::statement('UPDATE transaction_paid_leaves SET `days` = `days` - 1 WHERE `id` = '.$item_accept->paid_leave_id);
+                    DB::statement('UPDATE master_users SET `yearly_leave_remaining` = `yearly_leave_remaining` + 1 WHERE `id` = '.$item_accept->user_id);
+                    DB::table('accepted_paid_leaves')->where('id', '=', $item_accept->id)->delete();
+                }
             }
 
             $transaction_paid_leave = DB::table('transaction_paid_leaves')
-            ->whereIn('status', ['Diajukan', 'Pending'])->get();
+            ->whereIn('status', ['Diajukan', 'Pending'])
+            ->leftJoin('master_users','transaction_paid_leaves.user_id','=','master_users.id')
+            ->select(
+                'transaction_paid_leaves.*',
+                'master_users.division_id as division_id'
+                )
+            ->get();
             
             foreach ($transaction_paid_leave as $item_transaction) {
-                $transaction_interval = date_diff(date_create($item_transaction->paid_leave_date_start),date_create($item_transaction->paid_leave_date_end));
-                $transaction_total_day = $transaction_interval->format('%a') + 1;
+                if ($item_transaction->division_id != 5) {
+                    $transaction_interval = date_diff(date_create($item_transaction->paid_leave_date_start),date_create($item_transaction->paid_leave_date_end));
+                    $transaction_total_day = $transaction_interval->format('%a') + 1;
 
-                $transaction_date_start = date('Y-m-d', strtotime('-1 days', strtotime($item_transaction->paid_leave_date_start)));
-                for ($x = 0; $x <$transaction_total_day; $x++) {
-                    $check_transaction_days = date('Y-m-d', strtotime('+1 days', strtotime($transaction_date_start)));
-                    if ($check_transaction_days == $check_year.'-'.$month.'-'.$check_day) {
-                        DB::statement('UPDATE transaction_paid_leaves SET `days` = `days` - 1 WHERE `id` = '.$item_transaction->id);
+                    $transaction_date_start = date('Y-m-d', strtotime('-1 days', strtotime($item_transaction->paid_leave_date_start)));
+                    for ($x = 0; $x <$transaction_total_day; $x++) {
+                        $check_transaction_days = date('Y-m-d', strtotime('+1 days', strtotime($transaction_date_start)));
+                        if ($check_transaction_days == $check_year.'-'.$month.'-'.$check_day) {
+                            DB::statement('UPDATE transaction_paid_leaves SET `days` = `days` - 1 WHERE `id` = '.$item_transaction->id);
+                        }
+                        $transaction_date_start = $check_transaction_days;
                     }
-                    $transaction_date_start = $check_transaction_days;
                 }
             }
 
             $data_schedule = DB::table('master_job_schedules')
             ->where('month','=',$check_month)
             ->where('year','=',$check_year)
+            ->leftJoin('master_users','master_job_schedules.user_id','=','master_users.id')
+            ->select(
+                'master_job_schedules.*',
+                'master_users.division_id as division_id'
+                )
             ->get();
 
             foreach ($data_schedule as $item) {
-                $date_shift = 'shift_'.$check_day;
-                $total_hour = $item->total_hour;
-                $shift = $item->$date_shift;
-                $shift_hour = $total_hour - check_hour_shift($shift);
-                
-                DB::table('master_job_schedules')
-                ->where('id', '=', $item->id)
-                ->update(['shift_'.$check_day => 'Off', 'total_hour' => $shift_hour]);
-            }
+                if ($item->division_id != 5) {
+                    $date_shift = 'shift_'.$check_day;
+                    $total_hour = $item->total_hour;
+                    $shift = $item->$date_shift;
 
-            if ($i == 1) {
-                // die();
+                    if( $shift != 'Off' || $shift != 'Cuti') {
+                        $shift_hour = $total_hour - check_hour_shift($shift);
+                        DB::table('master_job_schedules')
+                        ->where('id', '=', $item->id)
+                        ->update(['shift_'.$check_day => 'Cuti', 'total_hour' => $shift_hour]);
+                    }
+                }
             }
 
             MasterHoliday::create([
@@ -140,8 +161,6 @@ class HolidayController extends Controller
 
             $date = $check_days;
         }
-
-        // die();
         
         Alert::success('Berhasil!', 'Hari libur baru telah ditambahkan!');
         return redirect('/admin/holiday');
